Renamed kube-batchd to kube-batch.
