+++ Makefile
@@ -2,7 +2,7 @@ BIN_DIR=_output/bin
 RELEASE_VER=v0.2
 
 kube-batch: init
-	go build -o ${BIN_DIR}/kube-batchd ./cmd/kube-batchd/
+	go build -o ${BIN_DIR}/kube-batch ./cmd/kube-batch/
 
 verify: generate-code
 	hack/verify-gofmt.sh
@@ -17,9 +17,9 @@ generate-code:
 	${BIN_DIR}/deepcopy-gen -i ./pkg/apis/scheduling/v1alpha1/ -O zz_generated.deepcopy
 
 images: kube-batch
-	cp ./_output/bin/kube-batchd ./deployment/
-	docker build ./deployment/ -f ./deployment/Dockerfile -t kubesigs/kube-batchd:${RELEASE_VER}
-	rm -f ./deployment/kube-batchd
+	cp ./_output/bin/kube-batch ./deployment/
+	docker build ./deployment/ -f ./deployment/Dockerfile -t kubesigs/kube-batch:${RELEASE_VER}
+	rm -f ./deployment/kube-batch
 
 run-test:
 	hack/make-rules/test.sh $(WHAT) $(TESTS)
