+++ deployment/kube-batch/values.yaml
@@ -1,4 +1,4 @@
-# Default values for kube-batchd.
+# Default values for kube-batch.
 # This is a YAML-formatted file.
 # Declare variables to be passed into your templates.
 replicaCount: 1
