+++ doc/images/kube-batch.png
