+++ example/job.yaml
@@ -19,7 +19,7 @@ spec:
           requests:
             cpu: "1"
       restartPolicy: Never
-      schedulerName: kube-batchd
+      schedulerName: kube-batch
 ---
 apiVersion: scheduling.incubator.k8s.io/v1alpha1
 kind: PodGroup
