+++ hack/.golint_failures
@@ -1,5 +1,5 @@
-cmd/kube-batchd/app
-cmd/kube-batchd/app/options
+cmd/kube-batch/app
+cmd/kube-batch/app/options
 pkg/apis/scheduling/v1alpha1
 pkg/apis/utils
 pkg/scheduler
